Add curly braces around variable references Minor fixes

+++ zip-content/inc/common.sh
@@ -301,7 +301,7 @@ delete_recursive_wildcard()
   for filename in "$@"; do
     if test -e "$filename"; then
       ui_debug "Deleting '$filename'...."
-      rm -rf "$filename" || ui_error "Failed to delete files/folders" 105
+      rm -rf "${filename:?}" || ui_error "Failed to delete files/folders" 105
     fi
   done
 }
